Add files via upload
Business problem 

In today's digital age, the spread of misinformation through online news articles has become a significant concern. For a news aggregator platform, ensuring the accuracy and reliability of the content is crucial to maintain credibility and trust among its users. The challenge lies in distinguishing between authentic news articles and misleading or fabricated content.

To address this issue, our business aims to implement a Fake News Detection system. The primary objective is to automatically classify news articles as either "Real" or "Fake" based on their content. This will empower the news platform to filter out potentially misleading information, providing users with a more trustworthy and reliable news feed.